requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pymongo
passlib[bcrypt]
pyjwt
requests
python-dotenv
